% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{annot}
\alias{annot}
\title{Annotate cell types for single cell RNA data}
\usage{
annot(
  data,
  db = "cellmarker",
  species = "Human",
  tissue = NULL,
  p_cut = 0.5,
  test = "GSEA",
  scoretype = "std"
)
}
\arguments{
\item{data}{Name of a data frame containing the markers' Entrez IDs, cluster, and expression scores; Marker genes should be sorted in each cluster.
Order of the columns should be gene, cluster and score.}

\item{db}{Name of the reference database; CellMarker database is used in default.}

\item{species}{Human or Mouse. Human in default.}

\item{tissue}{Tissue types can be specified when running the analysis. Length of tissue can be larger than 1. 
The possible tissues can be seen using `data(cellmarker_tissue)`, `data(clustermole_tissue)` and `data(panglao_tissue)`.}

\item{p_cut}{Cutoff of the P value for GSEA.}

\item{test}{"GSEA" or "fisher"; "GSEA" is used in default.}

\item{scoretype}{Argument used for GSEA. Default value is "std". If all scores are positive, then scoretype should be "pos".}
}
\description{
Annotate cell types for single cell RNA data
}
\examples{
data(gene_pbmc)
result <- annot(gene_pbmc, db="cellmarker", species="Human", 
tissue=c("Blood", "Peripheral blood", "Blood vessel",
"Umbilical cord blood", "Venous blood"), p_cut=0.3, test="GSEA", scoretype="pos")

}
