% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmarker_tissue.R
\docType{data}
\name{cellmarker_tissue}
\alias{cellmarker_tissue}
\title{Tissues in CellMarker database.}
\format{
A list with 2 elements:
\describe{
  \item{Human}{Human tissue}
  \item{Mouse}{Mouse tissue}
  
}
}
\usage{
cellmarker_tissue
}
\description{
A list containing 2 elements: Human tissues and Mouse tissues.
}
\keyword{datasets}
