% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermole_tissue.R
\docType{data}
\name{clustermole_tissue}
\alias{clustermole_tissue}
\title{Tissues in Clustermole database.}
\format{
A list with 2 elements:
\describe{
  \item{Human}{Human tissue}
  \item{Mouse}{Mouse tissue}
  
}
}
\usage{
clustermole_tissue
}
\description{
A list containing 2 elements: Human tissues and Mouse tissues.
}
\keyword{datasets}
