% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coremarkers.R
\name{coremarkers}
\alias{coremarkers}
\title{Title Summarize markers contirbuting to the cell type annotation}
\usage{
coremarkers(test, data, species)
}
\arguments{
\item{test}{Test used to annotation cell types: "GSEA" or "fisher"}

\item{data}{Annotation results.}

\item{species}{"Human" or "Mouse"}
}
\value{
A data frame containing genes contributed to cell annotation
}
\description{
Title Summarize markers contirbuting to the cell type annotation
}
\examples{
## core_markers <- coremarkers("GSEA", data)

}
