% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_pbmc.R
\docType{data}
\name{gene_pbmc}
\alias{gene_pbmc}
\title{Differential expressed marker genes in 9 clusters.}
\format{
A data frame with 727 rows and 3 variables:
\describe{
  \item{gene}{Entrez IDs of the marker genes}
  \item{cluster}{Cluster}
  \item{score}{Average of log 2 fold changes getting from the process procedure}
  
}
}
\source{
\url{https://cf.10xgenomics.com/samples/cell/pbmc3k/pbmc3k_filtered_gene_bc_matrices.tar.gz}
}
\usage{
data(gene_pbmc)
}
\description{
A data frame containing marker genes, clusters as well as the average of log 2 fold changes.
The original data set is from 10X genomics, and we followed the standard workflow 
provided by Seurat package to process data, and then format to get the data frame.
}
\keyword{datasets}
