% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsymbol.R
\name{mapsymbol}
\alias{mapsymbol}
\title{Title Convert gene symbol to Entrez ID}
\usage{
mapsymbol(d, species)
}
\arguments{
\item{d}{A data frame where first column contains gene symbols.}

\item{species}{"Human" or "Mouse".}
}
\value{
A data frame containing gene symbols and the corresponding Entrez ID
}
\description{
This function is used to convert the gene symbol to Entrez Id. Used in 
easyct function.
}
