% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panglao_tissue.R
\docType{data}
\name{panglao_tissue}
\alias{panglao_tissue}
\title{Tissues in Panglao database.}
\format{
A list with 2 elements:
\describe{
  \item{Human}{Human tissue}
  \item{Mouse}{Mouse tissue}
  
}
}
\usage{
data(panglao_tissue)
}
\description{
A list containing 2 elements: Human tissues and Mouse tissues.
}
\keyword{datasets}
