% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{plot_dot}
\alias{plot_dot}
\title{Create dot plot for annotation results}
\usage{
plot_dot(test = "GSEA", data)
}
\arguments{
\item{test}{Test used to annotate cell types: "GSEA" or "fisher"}

\item{data}{Annotation results}
}
\value{
Plots
}
\description{
Create dot plot for annotation results
}
\examples{
data(gene_pbmc)
result <- easyct(gene_pbmc, db="cellmarker", species="Human", 
tissue=c("Blood", "Peripheral blood", "Blood vessel",
"Umbilical cord blood", "Venous blood"), p_cut=0.3, test="GSEA", scoretype="pos")
plot_dot("GSEA", result)

}
