% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Title Annotate cell types for single cell RNA data}
\usage{
process_results(test, data)
}
\arguments{
\item{test}{Test used to annotation cell types: "GSEA" or "fisher"}

\item{data}{Annotation results.}
}
\value{
A data frame used to generate plots.
}
\description{
This function is used to process the annotation test results. 
Processed data will be used to generate plots.
}
