% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fisher.R
\name{test_fisher}
\alias{test_fisher}
\title{Fisher exact test used in function 'easyct'}
\usage{
test_fisher(testgenes, ref, cols)
}
\arguments{
\item{testgenes}{A data frame containing query genes and the expression scores.}

\item{ref}{The reference data base.}

\item{cols}{Column names of the input data frame}
}
\value{
A data frame containg the results of fisher's exact test.
}
\description{
This function is used to conduct the modified Fisher's exact test.
}
