% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{annot}
\alias{annot}
\title{Annotate cell types for single cell RNA data}
\usage{
annot(
  data,
  db = "cellmarker",
  species = "Human",
  tissue = NULL,
  p_cut = 0.5,
  test = "GSEA"
)
}
\arguments{
\item{data}{Name of a data frame containing the markers' Entrez IDs, cluster, and expression scores; Marker genes should be sorted in each cluster.
Order of the columns should be gene, cluster and score.}

\item{db}{Name of the reference database; CellMarker database is used in default.}

\item{species}{Human or Mouse. Human in default.}

\item{tissue}{Tissue types can be specified when running the analysis. Length of tissue can be larger than 1.}

\item{p_cut}{Cutoff of the P value for GSEA.}

\item{test}{Fisher exact test or GSEA; GSEA is used in default.}
}
\description{
Annotate cell types for single cell RNA data
}
