% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyct.R
\name{easyct}
\alias{easyct}
\title{We developed EasyCellType that can automatically examine the input marker lists 
obtained from existing software such as Seurat over the cell marker databases. 
Two quantification approaches to annotate cell types are provided: 
Gene set enrichment analysis (GSEA) and a fisher exact test combining with expression significance.}
\usage{
easyct(
  data,
  db = "cellmarker",
  species = "Human",
  tissue = NULL,
  p_cut = 0.5,
  test = "GSEA",
  scoretype = "std"
)
}
\arguments{
\item{data}{Name of a data frame containing the markers' Entrez IDs, cluster, 
and expression scores; Marker genes should be sorted in each cluster.
Order of the columns should be gene, cluster and score.}

\item{db}{Name of the reference database; CellMarker database is used in default.}

\item{species}{Human or Mouse. Human in default.}

\item{tissue}{Tissue types can be specified when running the analysis. Length of tissue can be larger than 1. 
The possible tissues can be seen using `data(cellmarker_tissue)`, `data(clustermole_tissue)` and `data(panglao_tissue)`.}

\item{p_cut}{Cutoff of the P value for GSEA.}

\item{test}{"GSEA" or "fisher"; "GSEA" is used in default.}

\item{scoretype}{Argument used for GSEA. Default value is "std". If all scores are positive, then scoretype should be "pos".}
}
\value{
List
}
\description{
We developed EasyCellType that can automatically examine the input marker lists 
obtained from existing software such as Seurat over the cell marker databases. 
Two quantification approaches to annotate cell types are provided: 
Gene set enrichment analysis (GSEA) and a fisher exact test combining with expression significance.
}
\examples{
data(gene_pbmc)
result <- easyct(gene_pbmc, db="cellmarker", species="Human", 
tissue=c("Blood", "Peripheral blood", "Blood vessel",
"Umbilical cord blood", "Venous blood"), p_cut=0.3, test="GSEA", scoretype="pos")

}
