% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Create bar plots for each cluster}
\usage{
plot_bar(test = "GSEA", data)
}
\arguments{
\item{test}{"GSEA" or "fisher"}

\item{data}{Annotation results}
}
\value{
Plot
}
\description{
Create bar plots for each cluster
}
\examples{
data(gene_pbmc)
result <- easyct(gene_pbmc, db="cellmarker", species="Human", 
tissue=c("Blood", "Peripheral blood", "Blood vessel",
"Umbilical cord blood", "Venous blood"), p_cut=0.3, test="GSEA", scoretype="pos")
plot_bar("GSEA", result)

}
