% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Create bar plots for each cluster}
\usage{
plot_bar(test = "GSEA", data)
}
\arguments{
\item{test}{"GSEA" or "fisher"}

\item{data}{Annotation results}
}
\description{
Create bar plots for each cluster
}
