% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{plot_dot}
\alias{plot_dot}
\title{Create dot plot for annotation results}
\usage{
plot_dot(test = "GSEA", data)
}
\arguments{
\item{test}{Test used to annotate cell types: "GSEA" or "fisher"}

\item{data}{Annotation results}
}
\description{
Create dot plot for annotation results
}
\examples{
plot_dot(test="GSEA", annot_results)

}
