% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Title Annotate cell types for single cell RNA data}
\usage{
process_results(test, data)
}
\arguments{
\item{test}{Test used to annotation cell types: "GSEA" or "fisher"}

\item{data}{Annotation results.}
}
\value{
data frame
}
\description{
Title Annotate cell types for single cell RNA data
}
